Contact model with memory
Fully functional implementation of the contact model with memory. Has been tested in sequential and parallel. The reload format has undergone significant changes and has been tested extensively as well.

Usage: the contact law name is "ERHContact", and the model parameters that need to be specified are: "stiff", "en", "muc", "ks", "mut" and "color". Note the new tangential stiffness "ks". If rolling friction is desired, the "mur" parameter is needed and is the rolling friction coefficient.

Important: this contact law introduces significant computational load compared to the ERContact law, so it can be preferable to use ERContact when all the system is constantly moving (i.e. a rotating drum setup). However, if even a sub-system is in a static configuration, this new contact law with history ERHContact will yield results much more physical.