Implement eul2lag & deal with periodic boundaries
Comments on changes:
* Implement the function eul2lag that interpolates the velocities from
the Eulerian mesh to the Lagrangian mesh, using a cosine-shaped
regularized Dirac. Tests seem successful for a complicated 2D flow field
in advect_caps.c... Although for the same time step no spatial
convergence is achieved. I suppose in this test the error in time
largely dominates that in space (there is convergence in time).

* Change the edges stuct: instead of copying the vertices, we simply
point to them.

* Deal with periodic boundary conditions in view-ft.h

* The lag2eul function is untested and probably bugged.

TODO:
* Deal with periodic boundary conditions to compute the edge length

* Test and debug the lag2eul function