Finish implementing neo-Hookean law
This version correctly reproduces the behaviour of a 2D capsule in a
shear flow: to compare our implementation to Breyiannis and Pozrikidis
(2000), we implement the linear elasticity law.

The variable viscosity solver is in the process of being implemented but
it is bugged for the moment.

TODO: the cleanup event leads to a segfault, so this should be fixed.