Implement bending force (qualitatively tested)
Implement a bending force in 2 dimensions. To this end, the computations
of normals of edges and nodes have been modified, and the computation of
the curvature at each node has been implemented and tested (on a circle:
we could test it on an ellipse as well). The curvature accuracy is
second-order with respect to the Lagrangian mesh.

The bending-2d.h header file uses the tools developed above to compute
the 2D bending force for a RBC. This force has been tested qualitatively
in a shear flow: we simply confirmed that the Taylor deformation
parameter decreases with increasing bending resistance.