Fix MPI bug in caps_viscosity
The grid-gradient is now a centered vector, it simplifies the
implementation and seemed to fix the MPI bug.

In reg-dirac.h, a function tag_ibm_stencils has been written to simplify
the adaptivity step, and not rely on the (too wide) scalar divG.

TODO: the code still crashed in MPI when using only one core. It should
be fixed.